Prevent getty on various services killing us (#979174)
Unfortunately, systemd doesn't allow to specify something like
'Before=getty.service serial-getty.service' so we must list all consoles and
serial-consoles manually. The proper solution would be to have a special
initial-setup.target, but that's a long-term goal that would require a lot more
work. This is a worse but small patch for the issue.