Read /etc/os-release to get product title (#1000426)
The /etc/os-release file contains, among the other things, the release name that
we want to display in the UI.

For GUI we can use just the product title because the string "INITIAL SETUP" is
shown in the top-left corner. For TUI we should inform user that this is an
initial setup of the system.

Also rename the hub files not to clash with the the Python package name.