Switch Initial Setup to Python 3 (#1244558)
Make Initial Setup use Python 3 by changing the startup script,
dependencies and by changing a single piece of Python 3 incompatible code.

Also remove usage of the python-di module - it has not yet been ported
to Python 3 and is not really needed. It has been apparently added for
use by tests, but there are no tests. It can of course be added back
once it is ported to Python 3 and once there are some tests that use it.