Make Initial Setup startup more robust (#1249598)
Use just a single primary systemd unit, named initial-setup.service, in place
of the two text and graphical units used previously. This should improve
robustness by placing less emphasis on systemd correctly resolving
unit dependencies.

The single unit file will run a shell script that checks if the GUI executable
exists and runs it if it does. If there is no GUI executable the TUI executable will be run.

For backward compatibility the old units (initial-setup-text.service and
initial-setup-graphical.service) are still provided and work like this:
* text enabled -> TUI will run (even if GUI is installed)
* graphical enabled -> GUI will run
* text & graphical enabled -> GUI will run

Note that the old units are deprecated and if possible initial-setup.service
should be used instead.

All scripts and executables have also been moved to /usr/libexec
to make it readily apparent they are not meant to be executed manually
during normal operation.

On the other hand it is now easier to debug Initial Setup startup issues
manually when one *does* run the executables by hand from the
/usr/libexec/initial-setup/ folder as log messages are now printed to stdout
and the usual --help and --version flags are supported.

Furthermore the core Initial Setup file has been refactored to be
easier to maintain, more docstrings have been added and the current
Initial Setup version is logged at start of the run.

Resolves: rhbz#1249598