Make Initial Setup startup more robust
Use just a single systemd service, named initial-setup.service, in place
of the two text and graphical services. This should improve
robustness by placing less emphasis on systemd correctly resolving
unit dependencies.

The single unit file will run a shell script that check if the GUI executable
exists and runs it if it does. If the is no GUI executable the TUI executable will be run.

All scripts and executables have also been moved to /usr/libexec
to make it readily apparent they are not meant to be executed manually
during normal operation.

On the other hand it is now easier to debug Initial Setup startup issues
manually when one *does* run the executables by hand from the
/usr/libexec/initial-setup/ folder as log messages are now printed to stdout
and the usual --help and --version flags are supported.

Furthermore the core Initial Setup file has been refactored to be
easier to maintain, more docstrings have been added and the current
Initial Setup version is logged at start of the run.