Start the window manager correctly (#1160891)
And also make sure the Anaconda/Initial Setup overrides
for Metacity are correctly applied.

The end result should be that Initial Setup should again
properly use the full screen real estate and that all
applications launched (such as the yelp help viewer) should
be properly decorated.

And also on Metacity the redundant top bar for the Initial Setup
window will be now hidden.