Make sure full screen is used & window header bar is hidden (#1290321)
Make sure the IS window makes use of the full available screen
by making sure a window manager is correctly started.

Also make use of a Metacity theme override to get rid of the
Window header bar so that the full available screen real estate
can be used.

Note that the override is only effective when using Metacity as a WM.
If another WM is used the header bar will be visible, but the window
will still be fullscreen.

Resolves: rhbz#1290321