Suppress an error message about missing initial-setup-graphical.unit (#1249598)
Even though we pass -q/--quiet to "systemctl is-enabled", it still
prints an error message to stderr if the given unit is missing.

As the initial-setup-gui package might not be always installed
the initial-setup-graphical.service might not be present, which
results in an annoying and superfluous error message in the console.

So redirect stderr to /dev/null for the checks to get rid of it.
We don't need to do the same thing for the initial-setup.service and
initial-setup-text.service as those are always available when the
initial-setup package is installed.

Related: rhbz#1249598