Suppress logging to stdout when TUI is started by s390 startup scripts
Initial setup TUI (and also GUI) log to stdout by default to help
with debugging (at the same time they still log to journal as usual).

This might not be desirable in all cases, such as when the TUI is
started from the s390 startup script.

So pass the --no-stdout-log option to the TUI executable to disable
stdout logging when the TUI is started from a s390 startup script.

Thanks to Jan Stodola for providing the patch!

Related: rhbz#1366776