Change how Initial Setup reconfig mode is started (#1342536)
Apparently Requires (Requires=initial-setup.service) is triggered
even if a Condition (ConditionPathExists=/.unconfigured) specified
by the service is not met (!).

So remove the Requires on the initial-setup.service and instead of
chain-starting the initial-setup.service (via Requires) directly run
the Initial Setup startup script (through reconfiguration-mode-enabled
script).

Also add some bits and pieces from the initial-setup.service unit file
to initial-setup-reconfigure.service so that they are as similar as
possible, as they are now effectively used in the same manner (to
directly run Initial Setup).

Related: rhbz#1342536z