Log unhandled exceptions to Journal (#1431227)
Make sure that unhandled exceptions in Initial Setup are logged
to the system Journal - IS already logs quite a lot of useful
stuff to Journal but unfortunately not the details about an
unhandled exception.

While ABRT might record the exception this requires ABRT to be running
and doesn't work if the Initial Setup files don't come from a package
(eq. during development/debugging). It will also be handy to have
the Initial Setup startup and status messages & traceback together in
the Journal. That way the traceback is also much more likely to end
ip in a bugreport & it should be almost certain to be grabbed by
sosreport.

So add a toplevel exception hook and log details about the exception to
the Journal. Also check if the "normal" Python-based logging machinery
has been initialized and log directly to Journal if not.

Resolves: rhbz#1431227