Run GUI on the s390 if possible (#1378082)
Run the Initial Setup GUI on the s390 if:
- the DISPLAY environmental variable is set
- either of initial-setup.service or initial-setup-graphical.service
  is enabled

This is achieved by executing the run-initial-setup script, which has
been extended to detect if $DISPLAY is set.

We can also remove the code for disabling the Initial Setup services
after successful run from the s390 startup scripts as the run-initial-setup
script already handles that.

Resolves: rhbz#1378082