Run the Initial Setup TUI on all usable consoles (#1438046)
This change attempts to address two issues with the Initial Setup TUI:

1) starting the TUI on a different console than the one a user
   connected to
2) starting the Initial Setup TUI on a console that is unusable.

Number 1) might generally not be a big issue on x86 PCs, where
Initial Setup is only used sporadically for simple post installation
configuration and where most users are simply using the graphical
console.

Things are much more serious for arm based single board
computers, where Initial Setup TUI more or less takes over the role of
an interactive installer, as the main installation method is image
based and the Initial Setup TUI running in reconfig mode provides
the only interactive configuration options.

Further users of arm devices make much more use of serial consoles, so
the x86 platform expectation that the user is mostly likely at the
graphical console does not hold.

Issue number 2) also has increased impact on arm devices, as arm boards
tend to have various esoteric serial consoles, some of which might not
be usable in an interactive mode. If such console is resolved as the
default one then the Initial Setup TUI might fail to open the console
and crash at startup.

The solution - run IS TUI on all usable TTYs

This should solve issue number 1) by providing access the Initial Setup
TUI regardles to what console (graphical, serial, etc.) the user is
using. It should also solve issue number 2) as consoles that can't be
opened or that fail during use are skipped.

The implementation - stdin and stdout redirection and multiplexing

The solution to issues 1) & 2) is achieved by redirecting the Initial
Setup TUI stdout and stdin to multiplexing code, which uses select() to:

- watch for any output and sends it to all usable consoles at once
- watch for input on any usable console and sends it to the
  Initial Setup stdin.

This efectively means that the Initial Setup TUI should be running on
all usable consoles and the user should be able to use any of the
consoles to interact with it.