Remove stdin & stdout definition from unit files (#1438046)
As initial setup now runs on all usable TTYs automatically, it no longer
needs systemd to provide it with a stdin and stdout.

This is actually a potential liability as the TTY might get resolved
to an unusable one, preventing Initial Setup from starting at all.

So remove the stdin and stdout from our service files and let the
multi-TTY handling code handle all of that.