Add support for password entry from arbitrary consoles (#1438046)
The getpass() function provided by the getpass module from the Python
standard library is used by the Anaconda/Initial Setup TUI for password
input. The getpass() function does this:

- provides a password prompt (eq. "Password:" or something similar)
- hides user input when user is typing the password

Unfortunately the default getpass() is hardcoded to expect user input
from /dev/tty, so it does not work with the recently added Initial Setup
multi-tty support.

To fix this a custom getpass() implementation (based on the default one)
has been added, which supports asking the user for password on an
arbitrary console. On the Anaconda side support has been added for using
custom getpass() implementation.

Like this users should be able to safely input passwords on any console
where the Initial Setup TUI can run.

Also bump the Anaconda version as we now need the Anaconda side
support for using a custom getpass() function.