Automate release creation
Add the moment one needs to call the "bumpver" target,
then manually create the release commit and then
call the "release" target, which tags the commit and
creates the tarball. This is pretty repetitive and
should be automated.

So add a new target called "commit" which creates a proper
release commit and make sure the release target
call the "bumpver" & "commit" targets before tagging and
creating the tarball.

This should turn a nominal Initial Setup release to typing:

make release

And everything will be done automatically in this order:
- bumpver
- release commit
- release tag
- tarball creation