Adjust to Hub behavior change
The continue_if_empty property has been added to graphical Hubs
to handle the case where the hub is empty during the Workstation
Live installation. While it makes sense for Anaconda to *not* autoquit
if the last hub is empty, it has always been the standard behavior
in Initial Setup. So override the default value for the Initial Setup GUI hub.