Add support for showing an EULA spoke
There are cases when it might be useful to show some sort of a license
or other text users should see before they start using the computer.

So add a new spoke that only shows up if such license text file is present.
Also if the file exists make sure users have to agree with it before
they can continue.

If no license text file is present the EULA spoke will not be shown
and Initial Setup will behave in a normal way.