Handle simpleline having an empty stack
During investigation of an apparent deadlock during some Initial Setup
runs (rhbz#1646568) we have found that the text UI library used by
Anaconda and Initial Setup will start it's mainloop and wait
indefinitely even if nothing is present on its screen stack.

To fix this simpleline will now by default raise an exception if starting the
main loop is attempted if no screens are scheduled to be run.

We need to catch this new exception and log a meaningful log message
to clearly indicate the UI is not starting as there is nothing to be
done.

Related: rhbz#1646568