Add CLI option to disable multi TTY handler
The multi TTY handler is an ideal solution for the
main Initial Setup usecase in text mode - to run
on all available consoles as we can never be sure
(due to the way consoles are implemented) which
console the user will use.

It might be the console running on the graphical
screen (if any, there might be no monitor connected
to the machine) or on one of the various serial
consoles.

The multi TTY handler solves this by running on
all available consoles, making Initial Setup TUI
accessible to the user regardless to what console
they choose to use.

Still, there is one use case multi TTY handler does
not cover - directly running Initial Setup in a
terminal session & this is the expected behavior
on the s390 (Initial Setup is started from a shell
profile script automatically upon SSH login).

So to handle this usecase, add the --no-multi-tty
command line option, that disables the multi tty
handler and just runs Initial Setup in the terminals
session on local stdout/stdin, as expected in this
case.

NOTE: It is actually possible to run a Initial Setup
instance with --no-multi-tty option while another
IS is running in the usual way with multi tty on the
regular graphical & serial consoles. This will work fine,
but you should make sure to disable the other
Initial Setup once the one you are interacting with has
been completed successfully.

(cherry picked from commit 6407d9765981f4cdf8fd56cd270d58f0331bca39)

Related: rhbz#1676439