Disable multi TTY handler when running in SSH session
On the s390 Initial Setup is started upon SSH login by
a profile script.

The multi TTY handler is not really intended to handle this
usecase, so disable it when starting Initial Setup in the
SSH session. This is achieved via the new --no-multi-tty
command line option, which disables multi TTY handler
when passed to Initial Setup at startup.

Also make sure to not only disable the Initial Setup service
but also to stop it upon successful completion of Initial Setup
via SSH. Otherwise the "regular" Initial Setup might still
be running and blocking login on graphical and serial consoles.

Resolves: rhbz#1676439