Remove old failing pre scriptlet
Back in RHEL 7 there was a case where the Initial Setup systemd service
could be running during system upgrade, resulting in an upgrade deadlock.

For that reason a %pre scriptlet has been added for the initial-setup
and initial-setup-gui packages, which made sure to disable the
Initial Setup services.

This is not needed anymore on RHEL8 as it is really not expected someone
will have the affected version of Initial Setup installed, running &
upgrading directly to the newest Initial Setup.

Also the necessary services the scriptlet checks for are not even there anymore,
as Initial Setup since then migrated to a combined service instead of two
services, covering the TUI and GUI separately. This results in nasty
error messages during Initial Setup package upgrades.

So lets remove the no longer needed %pre scripts, resulting in the
nasty error messages no longer being there.

Resolves: rhbz#1857180