Make sure the output from custom_getpass() is serialized after stdout
custom_getpass() writes the prompt to the file descriptor of the console where
the last input came from. At the same time, run() races with it, piping the
sys.stdout data to the same file descriptor (and others).

To make matters worse, the stdout data is buffered and with nobody flushing it
always loses the race, with important stuff being written out only after
custom_getpass() returns and somebody cares to flush it out:

  Password:
  Password (confirm):
  Password:
  Password (confirm):
  ================================================================================
  ================================================================================
  Root password

  Please select new root password. You will have to type it twice.

  The passwords you entered were different. Please try again.
  ================================================================================
  ================================================================================
  Root password

  Please select new root password. You will have to type it twice.

  The password is too short

This patch turns on line buffering, removing the necessity of explicit flushes
from the stdout writer side.

That alone wouldn't be sufficient because the stdout traffic could still be
delayed until the piper thread is awoken; though things wouldn't be mixed up
nearly as severely. To cope with this race the active console traffic is piped
to the piper thread as well and it is now responsible for ordering things.

It is still ugly but hey.