Fix translations for EULA spoke related strings
There were two strings that were not translated due to
wrong translation context - the Licensing category name
and the EULA spoke status message. In both cases the strings
were just marked for translation with N_() only to be translated
with _() from Anaconda with the "anaconda" context instead of
the correct context of "initial-setup" or translated with _()
imported from Anaconda.

For the category name, just switch to Initial Setup localization module
and translate the category name directly.

For the EULA spoke title, use workaround similar to one used
by the OSCAP addon - see bug 1673071 for more details.

Resolves: rhbz#1814952