Add TMT configuration to get Anaconda COPR build
Add just a smoke test for initial setup by importing the initial-setup python
module.

Initial setup depends on new version of Anaconda pretty often so let's use nightly
COPR builds to solve this correctly.

For local runs this test will just test sources but for remote PR it
should redefine ``path by the ``adjust`` part. This will change path to
somewhere else so the local ``initial_setup`` folder won't be used.

This adjust command is not implemented yet by TMT. It should be present in
1.2 version of TMT; expected to be released soon. Until then, it will work
locally but not on Packit. However, Packit for test run is broken even without
this change, so let's go with this solution even now.