Use `make test` to run TMT tests locally
This is a bit challenging because Gating & Packit install the build RPM file
automatically to the environment, however, that is not true for local test run.
Local run will be executed on sources. That will allow us to much quicker re-run
of the tests.

Implementation will save a lot of time during re-run of the tests. Basically we
will do provisioning and preparation for the tests just once. When you run
`make test` multiple times these will be skipped and only `discover` and
`execute` will be executed again. That means no installation of dependencies, no
download of container or restart of the container.

To update local container or fix other issues run `make test-cleanup` that will
remove cleanup last run so the next `make test` will again do a preparation.

Unfortunately, on Fedora 33 I'm hitting SELinux issue with test re-execution:
https://bugzilla.redhat.com/show_bug.cgi?id=1905968
To solve this please disable SELinux for the test re-execution.