Add support for using Mutter as window manager
Make it possible to run Initial Setup also with Mutter as the WM.

To avoid Mutter getting confused we need to tell it what session to use
(at the moment an X11 session as Initial Setup GUI still runs as root)
or else Initial Setup would not run full screen & there would be no
window decorations for application windows IS launches (yelp, connection
editor).

Also Mutter is apparently compatible with the same schemas/config files
as Metacity, so add these as well.

Thanks a lot to Jonas Adahl for the suggestion to set XDG_SESSION_TYPE,
which fixed all the issues! :)