Fix ownership of the gui folder (#1812463)
Fix ownership of the /usr/lib/python<version>/site-packages/initial_setup/gui
directory so it does not stay behind if the initial-setup-gui is uninstalled.

According to the RPM documentation it should be enough to mention the
full path without an asterisk at the end to both own the directory as
well as to include all the containing files.

Resolves: rhbz#1812463